requirements.txt
fastapi==0.92.0
notion_client==2.0.0
pydantic==1.10.1
python-dotenv==1.0.0
SQLAlchemy==2.0.4
